Added chat as a subscription and overhauled the authorization for other data models

See merge request mobile-ifi/msp/21ss/gruppe1!5

+++ backend/src/resolvers/task.ts
@@ -1,5 +1,7 @@
-import { ITask, IUser } from "../types/types";
+import { ITask, IUser, IEvent, ITaskManagement } from "../types/types";
 import Task from "../models/task";
+import TaskManagement from "../models/taskMgmt";
+import Event from "../models/event";
 import User from "../models/user";
 import mongoose from "mongoose";
 
@@ -10,7 +12,15 @@ export const resolver = {
 };
 
 export async function getTask(parent, args, context, info): Promise<ITask> {
-  return (await Task.findById(args.id)) as ITask;
+  // check authorization if user member of event for task
+  const task = (await Task.findById(args.id)) as ITask;
+  const taskMgmt = (await TaskManagement.findById(task.taskMgmtId)) as ITaskManagement;
+  const event = (await Event.findById(taskMgmt.eventId)) as IEvent;
+  if (!event.acceptedUsers.includes(context.req?.user?.id)) {
+    throw new Error("User is not member of event");
+  }
+
+  return task;
 }
 
 export async function getTasks(parent, args, context, info): Promise<ITask[]> {
@@ -24,27 +34,43 @@ export async function getTasks(parent, args, context, info): Promise<ITask[]> {
 }
 
 export async function createTask(parent, args, context, info): Promise<ITask> {
-  return new Promise<ITask>((resolve, reject) => {
-    const taskMgmtModel = new Task({
-      userId: args.userId ?? null,
-      taskMgmtId: args.taskMgmtId,
-      createdAt: Date.now(),
-      title: args.title,
-      description: args.description ?? null,
-      priority: args.priority ?? null,
-      done: false,
-    });
-    taskMgmtModel
-      .save()
-      .then((doc: ITask) => {
-        resolve(doc);
-      })
-      .catch((err) => reject(err));
+  // check authorization if user member of event for task
+  const task = (await Task.findById(args.id)) as ITask;
+  const taskMgmt = (await TaskManagement.findById(task.taskMgmtId)) as ITaskManagement;
+  const event = (await Event.findById(taskMgmt.eventId)) as IEvent;
+  if (!event.acceptedUsers.includes(context.req?.user?.id)) {
+    throw new Error("User is not member of event");
+  }
+
+  const taskMgmtModel = new Task({
+    userId: args.userId ?? null,
+    taskMgmtId: args.taskMgmtId,
+    createdAt: Date.now(),
+    title: args.title,
+    description: args.description ?? null,
+    priority: args.priority ?? null,
+    done: false,
   });
+  return taskMgmtModel
+    .save()
+    .then((doc: ITask) => {
+      return doc;
+    })
+    .catch((err) => {
+      throw new Error(err);
+    });
 }
 
 export async function updateTask(parent, args, context, info): Promise<ITask | null> {
   const task = (await Task.findById(args.id)) as ITask;
+
+  // check authorization if user member of event for task
+  const taskMgmt = (await TaskManagement.findById(task.taskMgmtId)) as ITaskManagement;
+  const event = (await Event.findById(taskMgmt.eventId)) as IEvent;
+  if (!event.acceptedUsers.includes(context.req?.user?.id)) {
+    throw new Error("User is not member of event");
+  }
+
   // update given args
   if (args.userId !== undefined) {
     task.userId = args.userId;
@@ -69,6 +95,14 @@ export async function updateTask(parent, args, context, info): Promise<ITask | n
 }
 
 export async function deleteTask(parent, args, context, info): Promise<string> {
+  // check authorization if user member of event for task
+  const task = (await Task.findById(args.id)) as ITask;
+  const taskMgmt = (await TaskManagement.findById(task.taskMgmtId)) as ITaskManagement;
+  const event = (await Event.findById(taskMgmt.eventId)) as IEvent;
+  if (!event.acceptedUsers.includes(context.req?.user?.id)) {
+    throw new Error("User is not member of event");
+  }
+
   return Task.findByIdAndRemove(args.id)
     .then(() => `Deleted task ${args.id}`)
     .catch((err) => {
