+++ backend/src/resolvers/taskMgmt.ts
@@ -16,7 +16,16 @@ export const resolver = {
 };
 
 export async function getTaskMgmt(parent, args, context, info): Promise<ITaskManagement> {
-  return (await TaskManagement.findById(args.id)) as ITaskManagement;
+  // check authorization if user member of event for task mgmt
+  const taskMgmt = (await TaskManagement.findById(args.id)) as ITaskManagement;
+  const event = (await Event.findById(taskMgmt.eventId)) as IEvent;
+
+  // check authorization
+  if (!event.acceptedUsers.includes(context.req?.user?.id)) {
+    throw new Error("User is not member of event");
+  }
+
+  return taskMgmt;
 }
 
 export async function getTaskMgmts(parent, args, context, info): Promise<ITaskManagement[]> {
@@ -30,21 +39,33 @@ export async function getTaskMgmts(parent, args, context, info): Promise<ITaskMa
 }
 
 export async function createTaskMgmt(parent, args, context, info): Promise<ITaskManagement> {
-  return new Promise<ITaskManagement>((resolve, reject) => {
-    const taskMgmtModel = new TaskManagement({
-      eventId: args.eventId,
-      createdAt: Date.now(),
-    });
-    taskMgmtModel
-      .save()
-      .then((doc: ITaskManagement) => {
-        resolve(doc);
-      })
-      .catch((err) => reject(err));
+  // check authorization
+  const event = (await Event.findById(args.eventId)) as IEvent;
+  if (!event.acceptedUsers.includes(context.req?.user?.id)) {
+    throw new Error("User is not member of event");
+  }
+
+  const taskMgmtModel = new TaskManagement({
+    eventId: args.eventId,
+    createdAt: Date.now(),
   });
+  return taskMgmtModel
+    .save()
+    .then((doc: ITaskManagement) => {
+      return doc;
+    })
+    .catch((err) => {
+      throw new Error(err);
+    });
 }
 
 export async function deleteTaskMgmt(parent, args, context, info): Promise<string> {
+  // check authorization
+  const event = (await Event.findById(args.eventId)) as IEvent;
+  if (!event.admins.includes(context.req?.user?.id)) {
+    throw new Error("User is not admin of event");
+  }
+
   // delete task management
   await TaskManagement.findByIdAndRemove(args.id)
     .then(() => `Deleted task management ${args.id}`)
